Anonymize the signed-off-by line example
Change the Signed-off-by line example to be anonymous. This should make the file
suitable for any other user's repo without needing to change the name to some
other than given. Another option was to use "Random J. Developer", which is also
a valid example. The problem I had with that, and the original, is that both
seem to imply that a real name is needed. For this repo, and any other I think I
will be making, real names are not needed, just some way to identify whom I'm
addressing when I have a question or issue with their commit. As long as the
email address is valid, and matches the given username in GitLab (or GitHub as
the case may be) the needs of adding the commit to the code are met.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>