Add license files to the repo
This repo is under the MIT license, as already placed in the README files.

The README files have been updated somewhat. The text version was slightly
reformated for line length. The MarkDown version was modified to make force
the license text into "code" formatting mode and the lines are set to be the
same as those in the text file.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>