Add files for dual-hosted repos
Having a repo on both GitLab and GitHub complicates the links for internal files
a bit. Referrencing the DCO.md file from the repo, for instance, on GitHub uses
a different URL format than it does on GitLab. With a repo on both hosts it's
still useful to referrence to "local" copy rather than the copy on the other.

There are other elements of the dual-hosting which can be complicating. The
files in the `dual-files` directory are templated to handle this, hopefully.

Made a few other changes to the original text of the CONTRIBUTING files, as well
as some clean-up in the README files.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>