Add template files for generic hosting
All links to internal files is removed as there is no way to know ahead of time
what URL pattern some other system might use. The repo could be on a self-hosted
GitWeb instance, or a remote private GitLab instance, or any number of other
arrangements. Including links to the named files in the CONTRIBUTING files is a
convenience for those reading them, and not critical.

As they are good resources, and the repo is not hosted on their competition,
most likely, the help page for generating a PGP key of both is still included.

Lastly, as an unknown hosting arrangement may, or may not have the Issues and
Pull request features of the big two, the sections dealing with issues, bugs,
and feture requests is completely removed.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>