Change code of conduct in contributing files
Removed the direct reference to No Code of Conduct from the files and made it a
reference to the `CODE_OF_CONDUCT` files in the repo itself. This will allow the
use of a different COC withouth replacing or editing the `CONTRIBUTING` files.

Also reordered section in the `README` files for a more logical flow.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>