Add simple components for the README building
I don't expect that most of the sections of the README files will be that much
different from one repo to the other, as templates anyway. Still, each section
will be it's own pair of files and controlled by settings in the config file.

Most of the sections are, from a template viewpoint, blank. The information on
Semantic Versioning and the copyright are standard enough to pre-build.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>