Change name of project to repo-template
The repo name is used in the sample data for the config file, so that was also
changed to the new name. Additionally, until I change my mind again, the launch
command file will be `git-start-repo`, usable as `git start-repo` as well.

Simplified the concepts for creating the files, removing the need to have unique
versions for different git hosts, such as GitHub, GitLab, or some other host, to
include Enterprise hosting and private hosting.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>